Obtain root account from everything
It was annoying to pass the root_account parameter everywhere. I
realized that we can obtain the root account from almost every object,
so do it that way. If we were dealing with multiple account structures,
this might be a problem, because we might pass objects from different
structures and accidentally work with different root accounts. However,
since we're not dealing with multiple account structures, it should be
okay.